// src/database/types/answer.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * user for the score app
 */
export interface Answer {
  /**
   * answer value
   */
  value: number;
  /**
   * The users ID
   */
  user: number;
  /**
   * The entry ID
   */
  entry: number;
  /**
   * the rank index for which this answer should be attributed
   */
  rank: number;
}