// src/database/types/user.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * user for the score app
 */
export interface User {
  /**
   * User ID
   */
  id?: number;
  /**
   * The users name
   */
  name: string;
  [k: string]: unknown;
}